+++ content/lights/spread.py
@@ -50,11 +50,9 @@ def update(rho, theta, photo, primary_color, secondary_color, led_count, strip):
     brightness = int(255 * (photo / 1024)) + 1
 
     # color of non-pixels
-    # bg_color = adjustBrightness(secondary_color, brightness)
     bg_color = secondary_color
 
     # color of spread by ball
-    # ball_color = adjustBrightness(primary_color, brightness)
     ball_color = primary_color
 
     # spread out the pixel color based on rho