+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 2019-08-26 1.10.32
+  - Light patterns
+  - Pattern brightness matches slider
+
 # 2019-08-07 1.10.31
   - Jimmy homing
   - Allow empty password on wifi
@@ -7,7 +11,6 @@
   - fixed sensored home on first homing (non-servos)
   - fix DR drift from deceleration
   - lock out home button when homing
-  - 
 
 # 2019-05-29 1.10.30
   - move sisbots over to webcenter